Rename files for consistency.
