+++ testArea.cpp
@@ -1,8 +1,8 @@
 #include <gtest/gtest.h>
 #include "max_area.h"
 #include "random_utils.h"
-#include "Point.h"
-#include "MergeSort.h"
+#include "point.h"
+#include "merge_sort.h"
 
 //sample compare function for ints
 int compare_by_ints(const void *p, const void *q)
