+++ testMerge.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <time.h>
 #include <stdlib.h>
-#include "MergeSort.h"
+#include "merge_sort.h"
 #include "random_utils.h"
